Added the initial config file
This CMAKE file was modified to comply with the newer versions of the UA toolkit (1.8.3) as well as the new versions of Quazar. Modifications include: Changing the root folder of the SDK from OPT (gave permission problem errors on Almalinux 9) to the root of the Quasar folder. Modified the header directories so that they follow the updated version of the UA sdk, and linked the libraries to the final executable.